bug is fixed
The bug is fixed. The details are as follows.

Try remove the threshold in the similarity in "postureMatrix_x ". I also
choose the score with the maximum in each word instead of summing them.
The result is 77% in 370 word in probe P50, where the feature of each
word has a dimension of 324.